// packages/smartsale-content/scripts/datocms/graphql/generated/cms/schema.ts
export type Scalars = {
  Boolean: boolean
  BooleanType: any
  CustomData: any
  DateTime: any
  Float: number
  FloatType: any
  Int: number
  IntType: any
  ItemId: any
  JsonField: any
  MetaTagAttributes: any
  String: string
  UploadId: any
}
// ...
export type ItemStatus = 'draft' | 'updated' | 'published'

export interface JsonFieldMultiLocaleField {
  locale: SiteLocale | null
  value: Scalars['JsonField'] | null
  __typename: 'JsonFieldMultiLocaleField'
}

/** Record of type Layout (layout) */
export interface LayoutRecord {
  _allAiFollowLinksLocales: JsonFieldMultiLocaleField[] | null
  _allAiResearchFollowLinksLocales: JsonFieldMultiLocaleField[] | null
  _allBackBitcashLocales: StringMultiLocaleField[] | null
  _allBackHomeLocales: StringMultiLocaleField[] | null
  _allBitcashNewsFollowLinksLocales: JsonFieldMultiLocaleField[] | null
  _allBitcoinFollowLinksLocales: JsonFieldMultiLocaleField[] | null
  _allCookieConsentCtaLocales: StringMultiLocaleField[] | null
  _allCookieConsentDescriptionLocales: StringMultiLocaleField[] | null
  _allCryptoFollowLinksLocales: JsonFieldMultiLocaleField[] | null
  _allEli5FollowLinksLocales: JsonFieldMultiLocaleField[] | null
  _allHomeFollowLinksLocales: JsonFieldMultiLocaleField[] | null
  _allInvestingFollowLinksLocales: JsonFieldMultiLocaleField[] | null
  _allNavigationCategoriesLocales: JsonFieldMultiLocaleField[] | null
  _allNavigationPoliciesTermsLocales: JsonFieldMultiLocaleField[] | null
  _allNavigationTopicLocales: JsonFieldMultiLocaleField[] | null
  _allSearchInputPlaceholderLocales: StringMultiLocaleField[] | null
  _allStartUpsFollowLinksLocales: JsonFieldMultiLocaleField[] | null
  _allSubscriptionCtaLocales: StringMultiLocaleField[] | null
  _allSubscriptionInputPlaceholderLocales: StringMultiLocaleField[] | null
  _allSubscriptionSubtitleLocales: StringMultiLocaleField[] | null
  _allSubscriptionTitleLocales: StringMultiLocaleField[] | null
  _createdAt: Scalars['DateTime']
  /** Editing URL */
  _editingUrl: Scalars['String'] | null
  _firstPublishedAt: Scalars['DateTime'] | null
  _isValid: Scalars['BooleanType']
  _modelApiKey: Scalars['String']
  _publicationScheduledAt: Scalars['DateTime'] | null
  _publishedAt: Scalars['DateTime'] | null
  /** Generates SEO and Social card meta tags to be used in your frontend */
  _seoMetaTags: Tag[]
  _status: ItemStatus
  _unpublishingScheduledAt: Scalars['DateTime'] | null
  _updatedAt: Scalars['DateTime']
  aiFollowLinks: Scalars['JsonField'] | null
  aiResearchFollowLinks: Scalars['JsonField'] | null
  backBitcash: Scalars['String'] | null
  backHome: Scalars['String'] | null
  bitcashNewsFollowLinks: Scalars['JsonField'] | null
  bitcoinFollowLinks: Scalars['JsonField'] | null
  cookieConsentCta: Scalars['String'] | null
  cookieConsentDescription: Scalars['String'] | null
  cryptoFollowLinks: Scalars['JsonField'] | null
  eli5FollowLinks: Scalars['JsonField'] | null
  homeFollowLinks: Scalars['JsonField'] | null
  id: Scalars['ItemId']
  investingFollowLinks: Scalars['JsonField'] | null
  navigationCategories: Scalars['JsonField'] | null
  navigationPoliciesTerms: Scalars['JsonField'] | null
  navigationTopic: Scalars['JsonField'] | null
  searchInputPlaceholder: Scalars['String'] | null
  startUpsFollowLinks: Scalars['JsonField'] | null
  subscriptionCta: Scalars['String'] | null
  subscriptionInputPlaceholder: Scalars['String'] | null
  subscriptionSubtitle: Scalars['String'] | null
  subscriptionTitle: Scalars['String'] | null
  __typename: 'LayoutRecord'
}
// ...
export type SiteLocale = 'en' | 'es' | 'zh' | 'id' | 'vi' | 'ko' | 'pt' | 'fr'

export interface StringMultiLocaleField {
  locale: SiteLocale | null
  value: Scalars['String'] | null
  __typename: 'StringMultiLocaleField'
}

export interface Tag {
  attributes: Scalars['MetaTagAttributes'] | null
  content: Scalars['String'] | null
  tag: Scalars['String']
  __typename: 'Tag'
}